+++ SourceCode/App/Validators/ExternalStationCustomerValidator.cs
@@ -1,0 +1,26 @@
+﻿using FluentValidation;
+using Microsoft.Extensions.Localization;
+using ModulesRegistry.Data;
+
+namespace ModulesRegistry.Validators
+{
+    public class ExternalStationCustomerValidator : AbstractValidator<ExternalStationCustomer>
+    {
+        public ExternalStationCustomerValidator(IStringLocalizer<App> localizer)
+        {
+            RuleFor(m => m.CustomerName)
+                .NotEmpty()
+                .MinimumLength(1)
+                .MaximumLength(50)
+                .MustBeOrdinaryText(localizer)
+                .MustBeCapitalizedCorrectly(localizer)
+                .WithName(n => localizer["Name"]);
+            RuleFor(m => m.OpenedYear)
+                .MustBeValidYear(localizer)
+                .WithName(n => localizer[nameof(n.OpenedYear)]);
+            RuleFor(m => m.ClosedYear)
+                .MustBeValidYear(localizer)
+                .WithName(n => localizer[nameof(n.ClosedYear)]);
+        }
+    }
+}