+++ SourceCode/Data/OperatingDay.cs
@@ -9,11 +9,11 @@ namespace ModulesRegistry.Data
     {
         public OperatingDay()
         {
-            CargoRelations = new HashSet<CargoRelation>();
-            ExternalStationCustomerCargos = new HashSet<ExternalStationCustomerCargo>();
+            //CargoRelations = new HashSet<CargoRelation>();
+            //ExternalStationCustomerCargos = new HashSet<ExternalStationCustomerCargo>();
             OperatingBasicDayBasicDays = new HashSet<OperatingBasicDay>();
             OperatingBasicDayOperatingDays = new HashSet<OperatingBasicDay>();
-            StationCustomerCargos = new HashSet<StationCustomerCargo>();
+            //StationCustomerCargos = new HashSet<StationCustomerCargo>();
         }
 
         public int Id { get; set; }
@@ -29,10 +29,10 @@ namespace ModulesRegistry.Data
         public bool IsSaturday { get; set; }
         public bool IsSunday { get; set; }
 
-        public virtual ICollection<CargoRelation> CargoRelations { get; set; }
-        public virtual ICollection<ExternalStationCustomerCargo> ExternalStationCustomerCargos { get; set; }
+        //public virtual ICollection<CargoRelation> CargoRelations { get; set; }
+        //public virtual ICollection<ExternalStationCustomerCargo> ExternalStationCustomerCargos { get; set; }
         public virtual ICollection<OperatingBasicDay> OperatingBasicDayBasicDays { get; set; }
         public virtual ICollection<OperatingBasicDay> OperatingBasicDayOperatingDays { get; set; }
-        public virtual ICollection<StationCustomerCargo> StationCustomerCargos { get; set; }
+        //public virtual ICollection<StationCustomerCargo> StationCustomerCargos { get; set; }
     }
 }