+++ SourceCode/Data/StationCustomerCargo.cs
@@ -22,7 +22,7 @@ namespace ModulesRegistry.Data
         public short? FromYear { get; set; }
         public short? UptoYear { get; set; }
 
-        //public virtual Cargo Cargo { get; set; }
+        public virtual Cargo? Cargo { get; set; }
         public virtual CargoDirection? Direction { get; set; }
         public virtual OperatingDay? OperatingDay { get; set; }
         public virtual CargoReadyTime? ReadyTime { get; set; }